[pytorch] Support process_group_agent "sending to itself"
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29253

Some operations can be simpler if a worker can send an rpc to itself.
The main reason for not doing previous was that Gloo doesn't support
self-sending.

That said, this changes the process_group_agent to skip the assert
check, and simply enqueue the rpc message in its receiving queue.
ghstack-source-id: 93474900

Differential Revision: [D18339715](https://our.internmc.facebook.com/intern/diff/D18339715/)