[jit] add `hasattr()`
Even though we're statically typed, this can be useful, e.g. as
shorthand when iterating through a module list.

Theoretically this could be dynamic--unlike `getattr` or something we
don't need static information about the attribute's type. It's more
general that way, but less consistent with our other stuff. If people
think it's better as a dynamic check then let me know.

ghstack-source-id: 4f8bf25a87cd3d8d85d463955fded1d7ccf4532e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29332